Make configuration handling on non-windows multi-level (Qt default ini)
The code currently uses two different configuration files, the Qt
native one to store the directory used in QPersistantFileDialog and
DInsightReportWindow, and a INI file named insight.conf for everything
else.  The first use is kept unchanged, while the latter is changed
for non-Windows builds.  Windows configuration handling is kept
unchanged to stay compatible with the already released version.  On
non-Windows platforms the code is changed to look in /usr/share/,
/etc/ and ~/.config/ for insight.ini instead of looking for
insight.conf only in the current working directory.

This bring the program more in line with Linux Filesystem Hierarcy
standard on Unix.

Note, the Qt QSettings documentation state that the native file to
look for is named insight.conf, but at least on Linux the file used by
QPersistantFileDialog and DInsightReportWindow is insightrc.  The file
used by DInsightConfig is insight.ini after this change, thus keeping
the two configuration systems in two separeate files.

This change partly addresses issue #9.